backend/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
python-doctr[pytorch]
sentence-transformers
faiss-cpu
scikit-learn
pandas
numpy
pydantic
groq
python-dotenv
pillow
opencv-python-headless
torch
torchvision
joblib
pyyaml
